+++ BenchManager/BenchLib/AppTyps.cs
@@ -24,14 +24,23 @@ namespace Mastersign.Bench
         /// <summary>The name of the app typ for Ruby gems.</summary>
         public const string RubyPackage = "ruby-package";
 
-        /// <summary>The name of the app typ for generic Python packages, managed by PIP.</summary>
+        /// <summary>The name of the app typ for generic Python packages from PyPI, managed by PIP.</summary>
         public const string PythonPackage = "python-package";
 
-        /// <summary>The name of the app typ for Python 2 packages, managed by PIP.</summary>
+        /// <summary>The name of the app typ for Python 2 packages from PyPI, managed by PIP.</summary>
         public const string Python2Package = "python2-package";
 
-        /// <summary>The name of the app typ for Python 3 packages, managed by PIP.</summary>
+        /// <summary>The name of the app typ for Python 3 packages from PyPI, managed by PIP.</summary>
         public const string Python3Package = "python3-package";
+
+        /// <summary>The name of the app typ for generic Python wheel files, managed by PIP.</summary>
+        public const string PythonWheel = "python-wheel";
+
+        /// <summary>The name of the app typ for Python 2 wheel files, managed by PIP.</summary>
+        public const string Python2Wheel = "python2-wheel";
+
+        /// <summary>The name of the app typ for Python 3 wheel files, managed by PIP.</summary>
+        public const string Python3Wheel = "python3-wheel";
 
         /// <summary>The name of the app typ for NuGet packages.</summary>
         public const string NuGetPackage = "nuget-package";